refactor: UserProject Module 분리
UserService에서 UserProjectRepository 사용을 위해 의존성 주입 중 의존성 순환 에러가 발생함.
의존도를 낮추고자 ProjectModule에서 UserProjectModule을 분리함.
