+++ src/modules/plans/services/plans.service.ts
@@ -140,14 +140,10 @@ export class PlansService {
     }
 
     // 일정 생성
-    async createPlan(
-        qr: QueryRunner,
-        projectId: number,
-        date: Date
-    ): Promise<CreatePlanResponse> {
+    async createPlan(qr: QueryRunner, projectId: number, date: Date): Promise<CreatePlanResponse> {
         // 유효한 식별자인지 & 사용자 권한 check
         const project = await this.projectsService.isProjectExists(projectId, qr.manager);
-        
+
         // 프로젝트 생성일자와 일정 생성일자 비교
         if (project.createdAt > date) {
             throw new PlanDateConflictException({
