+++ src/modules/projects/user-projects/user-project.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { UserProject } from './entities/user-projects.entity';
+import { UserProjectRepository } from './repositories/user-project.repository';
+
+@Module({
+    imports: [TypeOrmModule.forFeature([UserProject])],
+    providers: [UserProjectRepository],
+    exports: [UserProjectRepository],
+})
+export class UserProjectModule {}
