+++ src/modules/users/users.module.ts
@@ -5,11 +5,17 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/users.entity';
 import { UploadModule } from 'src/infra/upload/upload.module';
 import { MasterPortfoliosModule } from '../master-portfolios/master-portfolios.module';
-import { UserProject } from '../projects/entities/userProjects.entity';
 import { UserRepository } from './repositories/user.repository';
+import { UserProjectModule } from '../projects/user-projects/user-project.module';
+import { UserProject } from '../projects/user-projects/entities/user-projects.entity';
 
 @Module({
-    imports: [TypeOrmModule.forFeature([User, UserProject]), UploadModule, MasterPortfoliosModule],
+    imports: [
+        TypeOrmModule.forFeature([User, UserProject]),
+        UploadModule,
+        MasterPortfoliosModule,
+        UserProjectModule,
+    ],
     controllers: [UsersController],
     providers: [UsersService, UserRepository],
     exports: [UsersService, TypeOrmModule],
